proyecto_scomp/requirements.txt
streamlit
pandas
pdfplumber
fpdf2
google-generativeai
openpyxl
python-dotenv